Fix: bordered div in index.html no longer on map
This is a prettier alternative to the border extending all the way to the let's
connect area due to clashing of dark and yellow. Also added documentation to
helper.js added features (the line divs). Need to make sure that the border at
the top of #mapDiv appears on small displays, otherwise the .bordered border
will look horrible on them.